// src/api/models/calculo-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface CalculoResponse {
    'protocoloTitulo'?: string;
    'medicamentoNome'?: string;
    'referenciaBibliografica'?: string;
    'pesoConsideradoKg'?: number;
    'doseMinimaMg'?: number;
    'doseMaximaMg'?: number;
    'volumeMinimoMl'?: number;
    'volumeMaximoMl'?: number;
    'concentracaoUtilizada'?: string;
    'frequencia'?: string;
    'doseCalculadaMg'?: number;
    'volumeCalculadoMl'?: number;
    'via'?: string;
    'duracao'?: string;
    'alertas'?: Array<string>;
    'statusSeguranca'?: CalculoResponseStatusSegurancaEnum;
    'mensagemSeguranca'?: string;
    'refMin'?: number;
    'refMax'?: number;
    'refFonte'?: string;
}

export const CalculoResponseStatusSegurancaEnum = {
    Seguro: 'SEGURO',
    Subdose: 'SUBDOSE',
    Superdose: 'SUPERDOSE',
    SemReferencia: 'SEM_REFERENCIA',
    NaoValidado: 'NAO_VALIDADO'
} as const;

export type CalculoResponseStatusSegurancaEnum = typeof CalculoResponseStatusSegurancaEnum[keyof typeof CalculoResponseStatusSegurancaEnum];